+++ src/os2datascanner/projects/report/reportapp/views/utilities/ews_utilities.py
@@ -19,32 +19,41 @@ logger = structlog.get_logger("reportapp")
 
 
 def find_exchange_grant(org) -> (bool, EWSGrant | GraphGrant | str):  # noqa CCR001
-    # Try to get credentials, prefer GraphGrant if available (but prefer an
-    # EWSGrant /with/ a password over a GraphGrant /without/ a client secret)
+    # Try to get credentials, prefer EWSGrant if available (but prefer an
+    # GraphGrant /with/ a client secret over a EWSGrant /without/ a password)
+    # unless GraphGrant is prioritized by the organization
     grant: GraphGrant | EWSGrant | None = None
 
-    for candidate in GraphGrant.objects.filter(organization=org):
-        if candidate.client_secret:
-            if grant:
-                return (False, "too many credentials available")
-            else:
-                grant = candidate
-        else:
-            logger.warning(
-                    "skipping grant candidate with empty secret",
-                    candidate=candidate)
+    def pick_grant(grant_type, credential_attr):
+        for candidate in grant_type.objects.filter(organization=org):
 
-    if not grant:
-        for candidate in EWSGrant.objects.filter(organization=org):
-            if candidate.password:
+            nonlocal grant
+
+            if getattr(candidate, credential_attr):
                 if grant:
                     return (False, "too many credentials available")
                 else:
                     grant = candidate
             else:
                 logger.warning(
-                        "skipping grant candidate with empty secret",
-                        candidate=candidate)
+                        f"skipping grant candidate with empty {credential_attr}",
+                        candidate=candidate
+                )
+        return True, None
+
+    if org.prioritize_graphgrant:
+        valid, error = pick_grant(GraphGrant, "client_secret")
+        if not valid:
+            return False, error
+
+    else:
+        valid, error = pick_grant(EWSGrant, "password")
+        if not valid:
+            return False, error
+        if not grant:
+            valid, error = pick_grant(GraphGrant, "client_secret")
+            if not valid:
+                return False, error
 
     if grant:
         return (True, grant)